Move section tree code to dedicated module neurom._fst
Experimental module containing `load_neuron` function and
morphometrics modile `mm`.
