+++ examples/section_tree.py
@@ -1,17 +1,17 @@
 '''Build a section tree'''
 
 import sys
-from neurom.core import section_neuron as sn
+from neurom import _fst
 from neurom import ezy
 
 
 def do_new_stuff(filename):
     '''Use the section trees to get some basic stats'''
-    _n = sn.load_neuron(filename)
+    _n = _fst.load_neuron(filename)
 
-    n_sec = sn.n_sections(_n)
-    n_seg = sn.n_segments(_n)
-    sec_len = sn.get_section_lengths(_n)
+    n_sec = _fst.mm.n_sections(_n)
+    n_seg = _fst.mm.n_segments(_n)
+    sec_len = _fst.mm.section_lengths(_n)
 
     print 'number of sections:', n_sec
     print 'number of segments:', n_seg
