+++ neurom/_fst/tests/test_feature_compat.py
@@ -26,13 +26,13 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-'''Test neurom.features and section tree features compatibility'''
+'''Test neurom.features and neurom._fst features compatibility'''
 
 import os
 import numpy as np
 from nose import tools as nt
 from neurom.core.types import NeuriteType
-from neurom.core import section_neuron as sn
+from neurom import _fst
 from neurom.core.tree import i_chain2, ibifurcation_point
 from neurom.io.utils import load_neuron
 from neurom.features import get
@@ -84,103 +84,103 @@ class SectionTreeBase(object):
         nt.assert_equal(neurite_types, [n1.type for n1 in self.ref_nrn.neurites])
 
     def test_get_n_sections(self):
-        nt.assert_equal(sn.n_sections(self.sec_nrn), get('number_of_sections', self.ref_nrn)[0])
+        nt.assert_equal(_fst.mm.n_sections(self.sec_nrn), get('number_of_sections', self.ref_nrn)[0])
         for t in NeuriteType:
-            nt.assert_equal(sn.n_sections(self.sec_nrn, neurite_type=t),
+            nt.assert_equal(_fst.mm.n_sections(self.sec_nrn, neurite_type=t),
                             get('number_of_sections', self.ref_nrn, neurite_type=t)[0])
 
     def test_get_n_sections_per_neurite(self):
-        _equal(sn.get_n_sections_per_neurite(self.sec_nrn),
+        _equal(_fst.mm.n_sections_per_neurite(self.sec_nrn),
                get('number_of_sections_per_neurite', self.ref_nrn))
 
         for t in NeuriteType:
-            _equal(sn.get_n_sections_per_neurite(self.sec_nrn, neurite_type=t),
+            _equal(_fst.mm.n_sections_per_neurite(self.sec_nrn, neurite_type=t),
                    get('number_of_sections_per_neurite', self.ref_nrn, neurite_type=t))
 
     def test_get_n_segments(self):
-        nt.assert_equal(sn.n_segments(self.sec_nrn), get('number_of_segments', self.ref_nrn)[0])
+        nt.assert_equal(_fst.mm.n_segments(self.sec_nrn), get('number_of_segments', self.ref_nrn)[0])
         for t in NeuriteType:
-            nt.assert_equal(sn.n_segments(self.sec_nrn, neurite_type=t),
+            nt.assert_equal(_fst.mm.n_segments(self.sec_nrn, neurite_type=t),
                             get('number_of_segments', self.ref_nrn, neurite_type=t)[0])
 
     def test_get_number_of_neurites(self):
-        nt.assert_equal(sn.n_neurites(self.sec_nrn), get('number_of_neurites', self.ref_nrn)[0])
+        nt.assert_equal(_fst.mm.n_neurites(self.sec_nrn), get('number_of_neurites', self.ref_nrn)[0])
         for t in NeuriteType:
-            nt.assert_equal(sn.n_neurites(self.sec_nrn, neurite_type=t),
+            nt.assert_equal(_fst.mm.n_neurites(self.sec_nrn, neurite_type=t),
                             get('number_of_neurites', self.ref_nrn, neurite_type=t)[0])
 
     def test_get_section_lengths(self):
-        _close(sn.get_section_lengths(self.sec_nrn), get('section_lengths', self.ref_nrn))
+        _close(_fst.mm.section_lengths(self.sec_nrn), get('section_lengths', self.ref_nrn))
         for t in NeuriteType:
-            _close(sn.get_section_lengths(self.sec_nrn, neurite_type=t),
+            _close(_fst.mm.section_lengths(self.sec_nrn, neurite_type=t),
                    get('section_lengths', self.ref_nrn, neurite_type=t))
 
     def test_get_section_path_distances(self):
-        _close(sn.get_path_lengths(self.sec_nrn), get('section_path_distances', self.ref_nrn))
+        _close(_fst.mm.path_lengths(self.sec_nrn), get('section_path_distances', self.ref_nrn))
         for t in NeuriteType:
-            _close(sn.get_path_lengths(self.sec_nrn, neurite_type=t),
+            _close(_fst.mm.path_lengths(self.sec_nrn, neurite_type=t),
                    get('section_path_distances', self.ref_nrn, neurite_type=t))
 
-        pl = [sn.section_path_length(s) for s in i_chain2(self.sec_nrn.neurites)]
+        pl = [_fst.mm.section_path_length(s) for s in i_chain2(self.sec_nrn.neurites)]
         _close(pl, get('section_path_distances', self.ref_nrn))
 
     @nt.nottest
     def test_get_segment_lengths(self):
-        _equal(sn.get_segment_lengths(self.sec_nrn), get('segment_lengths', self.ref_nrn))
+        _equal(_fst.mm.segment_lengths(self.sec_nrn), get('segment_lengths', self.ref_nrn))
         for t in NeuriteType:
-            _equal(sn.get_segment_lengths(self.sec_nrn, neurite_type=t),
+            _equal(_fst.mm.segment_lengths(self.sec_nrn, neurite_type=t),
                    get('segment_lengths', self.ref_nrn, neurite_type=t))
 
     def test_get_soma_radius(self):
         nt.assert_equal(self.sec_nrn.soma.radius, get('soma_radii', self.ref_nrn)[0])
 
     def test_get_soma_surface_area(self):
-        nt.assert_equal(sn.soma_surface_area(self.sec_nrn), get('soma_surface_areas', self.ref_nrn)[0])
+        nt.assert_equal(_fst.mm.soma_surface_area(self.sec_nrn), get('soma_surface_areas', self.ref_nrn)[0])
 
     def test_get_local_bifurcation_angles(self):
-        _close(sn.get_local_bifurcation_angles(self.sec_nrn),
+        _close(_fst.mm.local_bifurcation_angles(self.sec_nrn),
                get('local_bifurcation_angles', self.ref_nrn))
 
         for t in NeuriteType:
-            _close(sn.get_local_bifurcation_angles(self.sec_nrn, neurite_type=t),
+            _close(_fst.mm.local_bifurcation_angles(self.sec_nrn, neurite_type=t),
                    get('local_bifurcation_angles', self.ref_nrn, neurite_type=t))
 
-        ba = [sn.local_bifurcation_angle(b)
+        ba = [_fst.mm.local_bifurcation_angle(b)
               for b in i_chain2(self.sec_nrn.neurites, iterator_type=ibifurcation_point)]
 
         _close(ba, get('local_bifurcation_angles', self.ref_nrn))
 
     def test_get_remote_bifurcation_angles(self):
-        _close(sn.get_remote_bifurcation_angles(self.sec_nrn),
+        _close(_fst.mm.remote_bifurcation_angles(self.sec_nrn),
                get('remote_bifurcation_angles', self.ref_nrn))
 
         for t in NeuriteType:
-            _close(sn.get_remote_bifurcation_angles(self.sec_nrn, neurite_type=t),
+            _close(_fst.mm.remote_bifurcation_angles(self.sec_nrn, neurite_type=t),
                    get('remote_bifurcation_angles', self.ref_nrn, neurite_type=t))
 
-        ba = [sn.remote_bifurcation_angle(b)
+        ba = [_fst.mm.remote_bifurcation_angle(b)
               for b in i_chain2(self.sec_nrn.neurites, iterator_type=ibifurcation_point)]
 
         _close(ba, get('remote_bifurcation_angles', self.ref_nrn))
 
     def test_get_section_radial_distances(self):
-        _close(sn.get_section_radial_distances(self.sec_nrn),
+        _close(_fst.mm.section_radial_distances(self.sec_nrn),
                get('section_radial_distances', self.ref_nrn))
 
         for t in NeuriteType:
-            _close(sn.get_section_radial_distances(self.sec_nrn, neurite_type=t),
+            _close(_fst.mm.section_radial_distances(self.sec_nrn, neurite_type=t),
                    get('section_radial_distances', self.ref_nrn, neurite_type=t))
 
     def test_get_trunk_origin_radii(self):
-        _equal(sn.get_trunk_origin_radii(self.sec_nrn), get('trunk_origin_radii', self.ref_nrn))
+        _equal(_fst.mm.trunk_origin_radii(self.sec_nrn), get('trunk_origin_radii', self.ref_nrn))
         for t in NeuriteType:
-            _equal(sn.get_trunk_origin_radii(self.sec_nrn, neurite_type=t),
+            _equal(_fst.mm.trunk_origin_radii(self.sec_nrn, neurite_type=t),
                    get('trunk_origin_radii', self.ref_nrn, neurite_type=t))
 
     def test_get_trunk_section_lengths(self):
-        _equal(sn.get_trunk_section_lengths(self.sec_nrn), get('trunk_section_lengths', self.ref_nrn))
+        _equal(_fst.mm.trunk_section_lengths(self.sec_nrn), get('trunk_section_lengths', self.ref_nrn))
         for t in NeuriteType:
-            _equal(sn.get_trunk_section_lengths(self.sec_nrn, neurite_type=t),
+            _equal(_fst.mm.trunk_section_lengths(self.sec_nrn, neurite_type=t),
                    get('trunk_section_lengths', self.ref_nrn, neurite_type=t))
 
 
@@ -188,14 +188,14 @@ class TestH5V1(SectionTreeBase):
 
     def setUp(self):
         super(TestH5V1, self).setUp()
-        self.sec_nrn = sn.load_neuron(os.path.join(H5V1_DATA_PATH, MORPH_FILENAME))
+        self.sec_nrn = _fst.load_neuron(os.path.join(H5V1_DATA_PATH, MORPH_FILENAME))
 
 
 class TestH5V2(SectionTreeBase):
 
     def setUp(self):
         super(TestH5V2, self).setUp()
-        self.sec_nrn = sn.load_neuron(os.path.join(H5V2_DATA_PATH, MORPH_FILENAME))
+        self.sec_nrn = _fst.load_neuron(os.path.join(H5V2_DATA_PATH, MORPH_FILENAME))
 
 
 class TestSWC(SectionTreeBase):
@@ -203,5 +203,5 @@ class TestSWC(SectionTreeBase):
     def setUp(self):
         self.ref_nrn = load_neuron(os.path.join(SWC_DATA_PATH, SWC_MORPH_FILENAME),
                                    mt.set_tree_type)
-        self.sec_nrn = sn.load_neuron(os.path.join(SWC_DATA_PATH, SWC_MORPH_FILENAME))
+        self.sec_nrn = _fst.load_neuron(os.path.join(SWC_DATA_PATH, SWC_MORPH_FILENAME))
         self.ref_types = [n.type for n in self.ref_nrn.neurites]
